fixes all CI issues except unit tests
